refactor: Sections 도메인 분리
