+++ src/test/java/study/unit/MockitoExtensionTest.java
@@ -29,7 +29,7 @@ public class MockitoExtensionTest {
     @Test
     void findAllLines() {
         // given
-        when(lineRepository.findAll()).thenReturn(Lists.newArrayList(new Line()));
+        when(lineRepository.findAll()).thenReturn(Lists.newArrayList(new Line("2호선", "green")));
         LineService lineService = new LineService(lineRepository, stationService);
 
         // when
